backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "flask==3.1.0",
    "flask-bcrypt==1.0.1",
    "flask-cors==5.0.1",
    "flask-jwt-extended==4.7.1",
    "flask-socketio==5.5.1",
    "mongoengine==0.29.1",
    "python-dotenv==1.0.1",
]
